fix(SDK): :bug: 修复 Go SDK 与 OpenAPI 的类型和参数对齐问题
- SessionService: AssistantMessage 移除 OpenAPI 不存在的 System 字段
- SessionService: UserMessage 补充缺失的 Agent/Model/Format/System/Tools 字段
- SessionService: CompactionPart 补充缺失的 tail_start_id 字段
- SessionService: 修复 Message.System 运行时类型注释 [[]string] → [string]
- SessionService: Init/Summarize/Command/Shell 方法参数标记 required
- ExperimentalService: Warp 参数补充 CopyChanges 字段, ID/Type 标记 required
- McpService: 修复 AuthAuthenticate/AuthRemove 未传递 query 参数的 bug
- PtyService: Pid 字段类型 int → int64（对齐 int64 惯例）

+++ event.go
@@ -3323,7 +3323,7 @@ type Pty struct {
 	Command string    `json:"command,required"`
 	Cwd     string    `json:"cwd,required"`
 	ID      string    `json:"id,required"`
-	Pid     int       `json:"pid,required"`
+	Pid     int64     `json:"pid,required"`
 	Status  PtyStatus `json:"status,required"`
 	Title   string    `json:"title,required"`
 	JSON    ptyJSON   `json:"-"`
